Refactor chagnes made on SpringConfiguration
